% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScorer.R
\name{buildScorer}
\alias{buildScorer}
\title{Build the Scorer}
\usage{
buildScorer(xdata, ydata, model, njob = 1, ...)
}
\arguments{
\item{xdata}{dataset to train. \code{data.frame} or \code{matrix} object with numeric columns.}

\item{ydata}{labels for the train. Two-level factor.}

\item{model}{type of model to use. One from "LR" or "RF".}

\item{njob}{number of nodes in parallel computing.}

\item{...}{further arguments to pass to \code{cv.glmnet\{glmnet\}} (only for LR).}

\item{ntree}{number of trees in the forest (only for RF)}

\item{cv.measure}{one from c("deviance","class","auc") (only for LR).}
}
\value{
The model, object of \code{cv.glmnet} or \code{randomForest} class
respectively to the passed \code{model} argument.
}
\description{
Build the scorer based on Logistic Regression (with LASSO penalty)
or Random Forest model.
}
\examples{
wmct <- WMCT(methPCBC, ENDOhealthy, G='SC', njob=3)
sign <- signatureWMCT(wmct, n=10)
train <- prepareTraining(methPCBC,ENDOhealthy,G='SC',signature = sign)
scorer <- buildScorer(train$X, train$Y, model="LR",
                      cv.measure='class', njob = 3,
                      intercept=TRUE, standardize=FALSE)
class(scorer)
plot(scorer)
}
\author{
Katarzyna Sobiczewska
}

