% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScorer_LRlasso.R
\name{buildScorer_LRlasso}
\alias{buildScorer_LRlasso}
\title{Logistic Regression (with LASSO penalty) Classifier}
\usage{
buildScorer_LRlasso(xdata, ydata, cv.measure, njob = 1, ...)
}
\arguments{
\item{xdata}{The training data. `data.frame` or `matrix` object with numeric columns.}

\item{ydata}{Labels for training. Two-level factor.}

\item{cv.measure}{one from c("deviance","class","auc").}

\item{njob}{number of nodes for parallel computing.}
}
\value{
Object of cv.glmnet class.
}
\description{
Model of Logistic Regression with LASSO penalty. \code{lambda} penalty
is fitted by cross-validation.
}
\author{
Katarzyna Sobiczewska
}

