% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScorer_RF.R
\name{buildScorer_RF}
\alias{buildScorer_RF}
\title{Random Forest Classifier}
\usage{
buildScorer_RF(xdata, ydata, ntree = 500, njob = 1)
}
\arguments{
\item{xdata}{The training data. `data.frame` or `matrix` object with numeric columns.}

\item{ydata}{Labels for training. Two-level factor.}

\item{ntree}{number of trees in the forest.}

\item{njob}{number of nodes for parallel computing.}
}
\value{
Object of class `randomForest`.
}
\description{
Building a classifier based on \code{randomForest} package with
possibility of parallel computing (prefered for large forests).
}
\author{
Katarzyna Sobiczewska
}

