% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoeffs.R
\name{getCoeffs}
\alias{getCoeffs}
\title{Extract Non-zero Model Coefficients}
\usage{
getCoeffs(model)
}
\arguments{
\item{model}{object created by \code{buildScorer} with argument \code{model="LR"}.}
}
\value{
Named vector of numerics. Names are predictors and values are
coefficients.
}
\description{
This function returns only non-zero coefficient values from model
created by \code{\link{buildScorer}} function with parameter
\code{model="LR"}.
}
\details{
LASSO penalty shrinks many coefficients to exactly \code{0}.
}
\examples{
wmct <- WMCT(methPCBC, ENDOhealthy, G='SC', njob=3)
sign <- signatureWMCT(wmct, n=10)
train <- prepareTraining(methPCBC,ENDOhealthy,G='SC',signature = sign)
scorer <- buildScorer(train$X, train$Y, model="LR",
                      cv.measure='class', njob = 3,
                      intercept=TRUE, standardize=FALSE)
getCoeffs(scorer)
}
\author{
Katarzyna Sobiczewska
}

