% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPCBC.R
\name{loadPCBC}
\alias{loadPCBC}
\title{Load PCBC}
\usage{
loadPCBC(path, data, meta_id = "UID", meta_class = "Diffname_short")
}
\arguments{
\item{fname}{name for target .csv data file and its metadata. Data file should be written
as \code{fname}.csv and metadata as meta_\code{fname}.csv.}
}
\value{
\code{ggplot} data.table object n x p, where n is for number of patients and p
for number of markers. First column of the object contains sample class (according to 
metadata \code{Diffname_short} information).
}
\description{
Load PCBC data from .csv files.
}
\author{
Katarzyna Sobiczewska
}

