% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonNAdata.R
\name{nonNAdata}
\alias{nonNAdata}
\title{Check NA values in given data frame}
\usage{
nonNAdata(data)
}
\arguments{
\item{xdata}{pis}

\item{ydata}{opis}
}
\value{
Logical vector of non NA rows.
}
\description{
Check for NA values in each row. Skip these rows and print appropriate message.
}
\details{
cos
}
\examples{
stay <- nonNA(xdata, ydata)
xdata <- xdata[stay,]
ydata <- ydata[stay]
}
\author{
Katarzyna Sobiczewska
}

