% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{perform}
\alias{perform}
\title{Perforn Classification}
\usage{
perform(data = "methylation", group = "meso", G = "SC",
  FUN = function(x) rank(x, na.last = "keep")/length(x),
  pcbc.dir = "~/RProjects/RStemnessScorer_backup/",
  meta_class = "Diffname_short", meta_id = "UID", nthreads = 3,
  ntrees = 2000)
}
\description{
Training and validation on stemness data and test on tumor data.
}

