% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoeffs.R
\name{plotCoeffs}
\alias{plotCoeffs}
\title{Plot the Coefficients}
\usage{
plotCoeffs(coeffs)
}
\arguments{
\item{coeffs}{list of coefficients estimated from different models: each element
should be formatted as named vector. If there is no need to put list as
one model was created you can pass just a named vector. Names are predictor
names (signature) and values are coefficients values estimated by the model.}
}
\value{
\code{ggplot} object that can be easily expanded and modified.
}
\description{
Compare finded coefficients extracted from Logistict Regression + LASSO
model(s) (see: \code{\link{buildScorer}(model=RF)}).
}
\examples{
wmct <- WMCT(methPCBC, ENDOhealthy, G='SC', njob=3)
sign <- signatureWMCT(wmct, n=10)
train <- prepareTraining(methPCBC,ENDOhealthy,G='SC',signature = sign)
measures <- c('class','auc','dev')
models <- lapply(measures, function(m)
                   buildScorer(train$X, train$Y, model="LR",
                               cv.measure=m, njob = 3))
names(models) <- measures
signatures <- lapply(models, getCoeffs)
(g <- plotCoeffs(signatures))
g + ggtitle("The title")

plotCoeffs(signatures[[3]])
}
\author{
Katarzyna Sobiczewska
}

