% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignature.R
\name{plotSignature}
\alias{plotSignature}
\title{Plot the Signature}
\usage{
plotSignature(x, y, signatures = NULL, horiz = TRUE)
}
\arguments{
\item{x}{dataset from which distributions should be visualised.
This object should contain all passed signatures.}

\item{y}{labels for each observations as a vector of characters or factors.}

\item{signatures}{list of signatures. One element of the list is a
signature from one model given as vector of characters. It is also
possible to pass one set signatures from a model as a single element. If
\code{signatures=NULL} than column names of \code{x} are used as a signature.}

\item{horiz}{logical; indicates if the boxplote should be horizontal.}
}
\value{
\code{\link{gridExtra}} object. Red color exposes signature
that is common for all passed models.
}
\description{
Visualize and compare distributions of selected signature between
different models.
}
\examples{
wmct <- WMCT(methPCBC, ENDOhealthy, G='SC', njob=3)
sign <- signatureWMCT(wmct, n=10)
train <- prepareTraining(methPCBC,ENDOhealthy,G='SC',signature = sign)
plotSignature(sign, train$X, train$Y)

## much more interesting
measures <- c('class','auc','dev')
models <- lapply(measures, function(m)
                   buildScorer(train$X, train$Y, model="LR",
                               cv.measure=m, njob = 3))
names(models) <- measures
signatures <- lapply(models, function(m) names(getCoeffs(m)))
plotSignature(signatures, train$X, train$Y)
plotSignature(signatures[-2], train$X, train$Y)
}
\author{
Katarzyna Sobiczewska
}

