% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImp.R
\name{plotVarImp}
\alias{plotVarImp}
\title{Plot of Variable Importance}
\usage{
plotVarImp(rf, n.var = 50, size = 0.5, labels = TRUE)
}
\arguments{
\item{rf}{the model class of \code{randomForest}.}

\item{n.var}{number of variables to display.}

\item{size}{point size.}

\item{labels}{logical; does the signature labels should be displayed?}
}
\value{
\code{ggplot} object that can be easily expanded and modified.
}
\description{
Dotchart of variable importance given by a \code{randomForest} model.
}
\examples{
wmct <- WMCT(methPCBC, ENDOhealthy, G='SC', njob=3)
sign <- signatureWMCT(wmct, n=100)
train <- prepareTraining(methPCBC,ENDOhealthy,G='SC',signature = sign)
rf <- buildScorer(train$X,train$Y, model="RF",ntree=500, njob = 3)
plotVarImp(rf, n.var=99, size=0.5)
}
\author{
Katarzyna Sobiczewska
}

