% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGATumorNormal.R
\name{RTCGATumorNormal}
\alias{RTCGATumorNormal}
\title{RTCGA Normal and Tumor}
\usage{
RTCGATumorNormal(data, group)
}
\arguments{
\item{data}{indicator of a RTCGA data set (as a string).}

\item{group}{name of one from ecto/endo/meso vectors.}
}
\value{
list of length 3. At the fist position there is vector of tumor samples, at the 
second: vector of normal samples. The last element of the list contains full data set 
where all samples across cancers in given \code{group} are concatenated.
}
\description{
The function collects group of cancers (given in 'cancerTypes/\code{group}.txt' files)
Splitting RTCGA.xxx data sets into tumor and normal patients respectively to 
barcode information.
}
\examples{

}
\author{
Katarzyna Sobiczewska
}

