% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R
\name{scorer}
\alias{scorer}
\title{Get a Scores}
\usage{
scorer(xtest, object, resp = 2)
}
\arguments{
\item{xtest}{dataset to test}

\item{object}{model created by \code{\link{buildScorer}}.}

\item{resp}{which class is the interesting response (respectively to \code{object$glmnet.fit$classnames} (\code{model="LR"}) or \code{object$classes} (\code{model="RF"}) output).}
}
\value{
Vector of scores for each observation. The score is a value
between 0 and 1.
}
\description{
Predict probabilities of membership to selected class based on given
model.
}
\author{
Katarzyna Sobiczewska
wmct <- WMCT(methPCBC, ENDOhealthy, G='SC', njob=3)
sign <- signatureWMCT(wmct, n=10)
train <- prepareTraining(methPCBC,ENDOhealthy,G='SC',signature = sign)
model <- buildScorer(train$X, train$Y, model="LR", cv.measure='dev')
probs <- scorer(train$X, model)
plot(sort(probs))
}

