% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_RF.R
\name{scorer_RF}
\alias{scorer_RF}
\title{Random Forest Scorer}
\usage{
scorer_RF(xtest, object, resp = 2)
}
\arguments{
\item{xtest}{dataset to test}

\item{object}{model created by \code{\link{buildScorer}(model='RF')}.}

\item{resp}{which from object$classes is the response.}
}
\value{
Vector of scores for each observation. The score is a value
between 0 and 1.
}
\description{
Random Forest Scorer
}
\author{
Katarzyna Sobiczewska
}

