% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureWMCT.R
\name{signatureWMCT}
\alias{signatureWMCT}
\title{Get the signature from Wilcoxon Multiple Comparisions Test}
\usage{
signatureWMCT(wmct, alpha = NULL, n = NULL)
}
\arguments{
\item{wmct}{the result of \code{\link{WMCT}} function.}

\item{alpha}{to get a signature for all adjusted p-values smaller than alpha.}

\item{n}{to get a signature for n smallest adjusted p-values.}
}
\value{
Vector of characters.
}
\description{
Set the cutoff on the adjusted p-values to obtain the interesting
 signature.
}
\details{
If both \code{n} and \code{alpha} values are passed,
than the \code{n} is ignored.
}
\examples{
wmctENDO <- WMCT(methPCBC, ENDOhealthy, G='SC')
signatureWMCT(wmctENDO, n=10)
}
\author{
Katarzyna Sobiczewska
}

