% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAscores.R, R/testTumor.R
\name{testTumor}
\alias{testTumor}
\title{Scores for Patients in TCGA Datasets.}
\usage{
testTumor(tcga, objects, resp = 2, verbose = TRUE)

testTumor(tcga, objects, resp = 2, verbose = TRUE)
}
\arguments{
\item{tcga}{dataset containing in the first column labels
given as factor.}

\item{objects}{list of one or more models to compare.}

\item{resp}{which class is the interesting response (respectively to
\code{object$glmnet.fit$classnames} (\code{model="LR"}) or
\code{object$classes} (\code{model="RF"}) output).}

\item{tcga}{dataset containing in the first column labels
given as factor.}

\item{objects}{list of one or more models to compare.}

\item{resp}{which class is the interesting response (respectively to
\code{object$glmnet.fit$classnames} (\code{model="LR"}) or
\code{object$classes} (\code{model="RF"}) output).}
}
\value{
\code{ggplot} object that can be easily expanded and modifed.

\code{ggplot} object that can be easily expanded and modifed.
}
\description{
Predict probabilities for all patients from TCGA dataset and get
a nice visualization.
Each patients should be assign to some group of observations, i.e. tumor
type. This function was made to compare results between each group of
patients.

Predict probabilities for all patients from TCGA dataset and get
a nice visualization.
Each patients should be assign to some group of observations, i.e. tumor
type. This function was made to compare results between each group of
patients.
}
\details{
If you prefer raw data rather than ggplot result use
\code{out$data} element. See example below to get more informations.

If you prefer raw data rather than ggplot result use
\code{out$data} element. See example below to get more informations.
}
\examples{
wmct_ENDO <- WMCT(methPCBC, ENDOhealthy, G='SC', njob=3)
signature_ENDO <- signatureWMCT(wmct_ENDO,alpha=0.05)
train <- prepareTraining(methPCBC, ENDOhealthy,G='SC',signature_ENDO)
measures <- c("class","auc","dev")
models <- lapply(measures, function(measure)
                 buildScorer(train$X,train$Y,model='LR',
                             measure, njob=3))
gg <- TCGAscores(ENDOtumor, models)
gg <- TCGAscores(ENDOtumor, models)
wmct_ENDO <- WMCT(methPCBC, ENDOhealthy, G='SC', njob=3)
signature_ENDO <- signatureWMCT(wmct_ENDO,alpha=0.05)
train <- prepareTraining(methPCBC, ENDOhealthy,G='SC',signature_ENDO)
measures <- c("class","auc","dev")
models <- lapply(measures, function(measure)
                 buildScorer(train$X,train$Y,model='LR',
                             measure, njob=3))
gg <- TCGAscores(ENDOtumor, models)
gg <- TCGAscores(ENDOtumor, models)
}
\author{
Katarzyna Sobiczewska

Katarzyna Sobiczewska
}

