% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WMCT.R
\name{WMCT}
\alias{WMCT}
\title{Wilcoxon Multiple Comparisions Test}
\usage{
WMCT(pcbc, tcga, G = "SC", adjust = "fdr", njob = 1)
}
\arguments{
\item{pcbc}{data set containing in the first column labels
given as factor. At least one level of the factor should be G. The
next columns are for predictors.}

\item{tcga}{tcga data set with healthy tissues.
The columns are predictors.}

\item{G}{label name for selected group of samples in \code{pcbc} dataset.}

\item{adjust}{method of correction. "none" if not interesting.
Another possible arguments are:
c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr")}

\item{njob}{number of clusters for parallel computing.}
}
\value{
Two-element list. The first element is a vector of adjusted
p.values obtained from multple comparisons test. The second one is a vector
of original p-values without any correction. You can use this element
to apply some other corretion method. See more: \code{\link{p.adjust}}.
}
\description{
Find siginificant differences between samples from
G class (PCBC) and healthy tissues (TCGA) + non G (PCBC)
for each predictor. To get interesting signature use
\code{\link{signatureWMCT}} function.
}
\details{
Two-sample Wilcoxon test is used for each of predictors
(for more details see: \code{\link{wilcox.test}}). As an alternative "two.sided"
was choosen. Note, that there are as many comparisions as many predictors
is on the overlap of `pcbc` and `tcga` sets. This approach requires
to use some of correction methods. Read more on \href{https://en.wikipedia.org/wiki/Multiple_comparisons_problem}{wikipedia}.
}
\examples{
wmct <- WMCT(methPCBC, ENDOhealthy, G='SC', njob=3)
wmct2 <- p.adjust(wmct[[2]],'bonferroni')
signatureWMCT(wmct, n=10)
}
\author{
Katarzyna Sobiczewska
}

